refactor(cc-addon-info): refactor api calls and create a `cc-addon-header.api.js` file
